simulate_population

Simulate population level data

simulate_population(
  data_structure,
  N,
  parameters,
  N_response = 1,
  known_predictors,
  model,
  family = "gaussian",
  link = "identity",
  pedigree,
  pedigree_type,
  phylogeny,
  phylogeny_type,
  cov_str,
  N_pop = 1
)

Arguments

data_structure

A matrix or dataframe with a named column for each grouping factor, including the levels

N

Sample size when data_structure is not specified

parameters

A list of parameters for each hierarchical level. See details.

N_response

The number of response variables, defaults to 1.

known_predictors

This option provides a way of inputting existing predictor variables, without simulating all predictors. This argument takes a list, with item 'predictors' and 'betas', where 'predictors' is a matrix or dataframe

model

Optional.

family

A description of the error distribution. Default "gaussian".

link

A description of the link function distribution. Default "identity".

pedigree

A list of pedigrees for each hierarchical level. Each pedigree must be matrix or data.frame, that is at least 3 columns, which correspond to ID, dam and sire.

pedigree_type

A list describing what kind of genetic variance is to be simulated from each pedigree. Default is 'additive', other options are 'dominance' and 'epistatic'. Makes use of relationship matrices created by the nadiv package.

phylogeny

A list of phylogenies for each hierarchical level. Each pedigree should be phylo class.

phylogeny_type

A list describing what mode of evolution should be simulated from each phylogeny. Options are 'brownian'(default) or 'OU'.

cov_str

A list of covariance structures for each hierarchical level.

N_pop

Number of populations. Default = 1

Value

Details

Parameter list ...

Examples